refactor(core): hoist the extract one-of + Format + derive_target into alint-core
Pure behavior-preserving relocation so a forthcoming core-side predicate
can reach the structured-document extract machinery (ScopeFilter already
lives in core; the extractor lived in alint-rules, dep runs rules->core).
No logic changes; the schema, all tests, and clippy stay green.

Moved alint-rules -> alint-core (pub(crate) -> pub, re-exported at the
crate root next to scope_filter):

- structured_format.rs (new): `Format` enum + `parse`/`label`/
  `detect_from_path`, plus the private helpers it needs (`strip_jsonc`,
  `MAX_XML_DEPTH`, `xml_depth_within_limit`, `xml_to_value`,
  `element_to_value`). `Format::parse`'s `alint_core::yaml_depth::*`
  references become `crate::yaml_depth::*`. `MAX_XML_DEPTH` is pub so the
  rule-level depth test in structured_path.rs can size its fixture.
- extract.rs (moved whole): `Extract`, `ExtractSpec`, `LinesOpts`,
  `WholeFileOpts`, `resolve`, `extract_values`, `is_non_literal`, and the
  private `structured` helper; `use crate::structured_path::Format` ->
  `use crate::structured_format::Format`. Its unit tests moved with it.
- pathsafe.rs (new in core): `normalize_confined` + its private
  `is_confined` helper, the Kani `model`/`kani_proofs`, and the
  normalize-specific unit/property tests. alint-rules/pathsafe.rs keeps
  the filesystem layer (confine/confine_read/resolved_within_root/
  out_of_root_note/Confined) and re-exports `normalize_confined` via
  `pub(crate) use alint_core::normalize_confined` so every existing
  `crate::pathsafe::normalize_confined` call site resolves unchanged.
- new `pub fn derive_target(from, to, path) -> Option<PathBuf>` in core
  pathsafe.rs, factored from file_graph's node_targets inline logic;
  file_graph now calls it (re-checking `from.captures` on `None` to keep
  the exact out-of-repo-target escape violation).

Plumbing:
- alint-core Cargo.toml gains toml/serde_json_path/roxmltree (for Format)
  and schemars (ExtractSpec/LinesOpts/WholeFileOpts derive JsonSchema),
  plus proptest as a dev-dep (moved property tests).
- alint-core/src/lib.rs declares mod extract/pathsafe/structured_format
  and re-exports Format, MAX_XML_DEPTH, Extract, ExtractSpec, LinesOpts,
  WholeFileOpts, extract_values, is_non_literal, normalize_confined,
  derive_target.
- importers (cross_file, registry_paths_resolve, file_graph,
  json_schema_passes) now `use alint_core::{...}`; alint-rules extract.rs
  deleted and its `mod extract;` removed.

migrated_option_schemas still emits a byte-identical extract_spec (now
referencing alint_core::ExtractSpec); gen-schema --check stays green.
