+++ crates/alint-core/Cargo.toml
@@ -23,9 +23,17 @@ ignore.workspace = true
 rayon.workspace = true
 regex.workspace = true
 tracing.workspace = true
+# Structured-document extract machinery hoisted from alint-rules:
+# TOML/XML/JSONPath parsing behind `Format`/`ExtractSpec`, whose
+# schemars derives feed `xtask gen-schema`.
+schemars.workspace = true
+serde_json_path.workspace = true
+toml.workspace = true
+roxmltree.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
+proptest.workspace = true
 
 [lints]
 workspace = true
