+++ crates/alint-core/src/lib.rs
@@ -9,15 +9,18 @@ mod config;
 pub mod did_you_mean;
 mod engine;
 mod error;
+mod extract;
 pub mod facts;
 pub mod git;
 pub mod jsonpath_diagnostics;
 mod level;
+mod pathsafe;
 mod registry;
 mod report;
 mod rule;
 mod scope;
 mod scope_filter;
+mod structured_format;
 pub mod template;
 mod walker;
 pub mod when;
@@ -34,8 +37,10 @@ pub use config::{
 };
 pub use engine::{Engine, RuleEntry};
 pub use error::{Error, Result};
+pub use extract::{Extract, ExtractSpec, LinesOpts, WholeFileOpts, extract_values, is_non_literal};
 pub use facts::{FactKind, FactSpec, FactValue, FactValues, evaluate_facts};
 pub use level::Level;
+pub use pathsafe::{derive_target, normalize_confined};
 pub use registry::{RuleBuilder, RuleRegistry};
 pub use report::{FixItem, FixReport, FixRuleResult, FixStatus, Report};
 pub use rule::{
@@ -47,5 +52,6 @@ pub use scope_filter::{
     ScopeFilter, ScopeFilterSpec, reject_scope_filter_on_cross_file,
     reject_scope_filter_with_reason,
 };
+pub use structured_format::{Format, MAX_XML_DEPTH};
 pub use walker::{FileEntry, FileIndex, MAX_ANALYZE_BYTES, WalkOptions, read_capped_or_skip, walk};
 pub use when::{WhenEnv, WhenError, WhenExpr};
